Move DB away from the project
