+++ crates/project/Cargo.toml
@@ -40,7 +40,6 @@ clock.workspace = true
 collections.workspace = true
 context_server.workspace = true
 dap.workspace = true
-db.workspace = true
 extension.workspace = true
 fancy-regex.workspace = true
 fs.workspace = true
